Implemented feature request for Bug #1641. Munge Process priority is now specified by the order of munge processes in the munge process folder. This is done from the munge process group editor now, by selecting the munge process in the table, and using the arrow buttons at the right side to move them up and down the list. The match engine will process the munge processes in this order. As such, the JSpinner in the munge process editor is gone. A few other changes were made:
* The match priority is now kept as an Integer, instead of a Short, since the munge process' index in the munge process folder is returned as an int. Fortunately, the match percent column in the repository schema should be large enough so no schema changes should be necessary.

* The Hibernate configuration for Project returns Munge Processes sorted by the match priority, so when munge processes are loaded with a project, they are placed in proper order in the folder.

* MatchMakerFolder passes an isCompound flag when firing an addChild event to prevent one of the listeners in MatchMakerSwingSession (SelectRemovedEditorListener? or something like that), from selecting the moved munge process editor when simply moving munge processes around, since this makes change the order of the munge processes really annoying.

+++ regress/ca/sqlpower/matchmaker/MatchPoolTest.java
@@ -307,8 +307,8 @@ public class MatchPoolTest extends TestCase {
 		insertSourceTableRecord(con, "2");
 		insertResultTableRecord(con, "1", "2", 15, "Munge_Process_One", PotentialMatchRecord.MatchType.UNMATCH);
 		pool.findAll(null);
-		mungeProcessOne.setMatchPriority(Short.valueOf("2"));
-		mungeProcessTwo.setMatchPriority(Short.valueOf("1"));
+		mungeProcessOne.setMatchPriority(Integer.valueOf("2"));
+		mungeProcessTwo.setMatchPriority(Integer.valueOf("1"));
 		
 		List<Object> keyList = new ArrayList<Object>();
 		keyList.add("1");
